+++ test/field.object.spec.js
@@ -28,42 +28,83 @@ describe('field.object', () => {
     // must force otherwise the data is not parsed
     emptyAllow: false,
   });
-
-  describe('general', () => {
-    logger.clear();
-    it('type lookup', async () => {
-      let r = await f.convert('object', {
-        type: 'John',
-        typeGuid: 'TST_CODE'
-      }, logger);
-      assert.equal(r.typeId, 123, 'translate the type');
-    });
-    it('type lookup, no find', async() => {
-      let r = await f.convert('object', {
-        type_: 'John',
-        typeGuid: 'TST_CODE'
-      }, logger);
-      assert.equal(r.typeId, 456, 'translate the type');
-    });
-    it('type lookup, no find', async() => {
-      let r = await f.convert('object', {
-        type_: 'John',
-        type: 'Jack',               // type is overloaded in removed because type_ does exist
-        typeGuid: 'TST_CODE'
-      }, logger);
-      assert.equal(r.typeId, 456, 'translate the type');
-    })
-
-  });
+  //
+  // describe('general', () => {
+  //   logger.clear();
+  //   it('type lookup', async () => {
+  //     let r = await f.convert('object', {
+  //       type: 'John',
+  //       typeGuid: 'TST_CODE'
+  //     }, logger);
+  //     assert.equal(r.typeId, 123, 'translate the type');
+  //   });
+  //   it('type lookup, no find', async() => {
+  //     let r = await f.convert('object', {
+  //       type_: 'John',
+  //       typeGuid: 'TST_CODE'
+  //     }, logger);
+  //     assert.equal(r.typeId, 456, 'translate the type');
+  //   });
+  //   it('type lookup, no find', async() => {
+  //     let r = await f.convert('object', {
+  //       type_: 'John',
+  //       type: 'Jack',               // type is overloaded in removed because type_ does exist
+  //       typeGuid: 'TST_CODE'
+  //     }, logger);
+  //     assert.equal(r.typeId, 456, 'translate the type');
+  //   })
+  //
+  // });
 
   describe('typeIsInsertOnly', () => {
-    logger.clear();
-    it('type lookup', async () => {
-      let r = await f.convert('object', {
+    class LookupInsertOnly extends Lookup {
+      async code(fieldName, def) {
+        let result;
+        if (!def.hasOwnProperty('fieldTypeInsertOnly')) {
+          result = -1 // it should ALWAYS have this property
+        } else if (def.fieldTypeInsertOnly) {
+          result = 1
+        } else {
+          result = 0;
+        }
+        return Promise.resolve(result);
+      }
+    }
+
+    // it('type lookup', async () => {
+    //   logger.clear();
+    //   let r = await f.convert('object', {
+    //     type: 'John',
+    //     typeInsertOnly: 1
+    //   }, logger);
+    //   assert.equal(r.typeInsertOnly, true, 'translate the type');
+    // });
+
+    it('lookup with the value', async() => {
+      logger.clear();
+      let field = new FieldObject({
+        lookupFunctionName: 'code',
+        lookup: new LookupInsertOnly(),
+        // must force otherwise the data is not parsed
+        emptyAllow: false,
+      });
+      let r = await field.convert('object', {
         type: 'John',
-        typeInsertOnly: 1
+        typeInsertOnly: 1,
+        typeGuid: 'NoRealy'
       }, logger);
-      assert.equal(r.typeInsertOnly, true, 'translate the type');
+      assert.equal(r.typeId, 1)
+      r = await field.convert('object', {
+        type: 'John',
+        typeInsertOnly: 0,
+        typeGuid: 'NoRealy'
+      }, logger);
+      assert.equal(r.typeId, 0);
+      r = await field.convert('object', {
+        type: 'John',
+        typeGuid: 'NoRealy'
+      }, logger);
+      assert.equal(r.typeId, 0)
     })
 
   })